+++ bmc-common/src/main/java/com/oracle/bmc/http/internal/ContentLengthFilter.java
@@ -8,6 +8,7 @@ import org.apache.commons.lang3.StringUtils;
 
 import javax.ws.rs.client.ClientRequestContext;
 import javax.ws.rs.client.ClientRequestFilter;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MultivaluedMap;
 
 /**
@@ -16,6 +17,17 @@ import javax.ws.rs.core.MultivaluedMap;
  */
 @Slf4j
 public class ContentLengthFilter implements ClientRequestFilter {
+    private final boolean removeZeroLengthHeader;
+
+    public ContentLengthFilter() {
+        // by default we remove the header regardless of value
+        this(true);
+    }
+
+    public ContentLengthFilter(boolean removeZeroLengthHeader) {
+        this.removeZeroLengthHeader = removeZeroLengthHeader;
+    }
+
     @Override
     public void filter(ClientRequestContext requestContext) {
         final MultivaluedMap<String, Object> headers = requestContext.getHeaders();
@@ -30,13 +42,19 @@ public class ContentLengthFilter implements ClientRequestFilter {
 
         String contentLengthHeader = null;
         for (String key : headers.keySet()) {
-            if (StringUtils.equalsIgnoreCase("content-length", key)) {
+            if (StringUtils.equalsIgnoreCase(HttpHeaders.CONTENT_LENGTH, key)) {
                 contentLengthHeader = key;
             }
         }
 
         if (contentLengthHeader == null) {
             LOG.debug("content-length not found for Method [{}], URI [{}]", method, uri);
+            return;
+        }
+
+        Object contentLengthValue = headers.getFirst(contentLengthHeader);
+        if (!removeZeroLengthHeader && "0".equals(contentLengthValue)) {
+            LOG.debug("Not removing zero content-length for Mehtod [{}], URI [{}]", method, uri);
             return;
         }