+++ bmc-core/src/main/java/com/oracle/bmc/core/ComputeAsync.java
@@ -272,7 +272,7 @@ public interface ComputeAsync extends AutoCloseable {
 
     /**
      * Creates a new console connection to the specified instance.
-     * Once the console connection has been created and is available,
+     * After the console connection has been created and is available,
      * you connect to the console using SSH.
      * <p>
      * For more information about console access, see [Accessing the Console](https://docs.cloud.oracle.com/Content/Compute/References/serialconsole.htm).