+++ bmc-core/src/main/java/com/oracle/bmc/core/ComputeManagement.java
@@ -47,7 +47,12 @@ public interface ComputeManagement extends AutoCloseable {
     AttachLoadBalancerResponse attachLoadBalancer(AttachLoadBalancerRequest request);
 
     /**
-     * Change the compartment of a cluster network.
+     * Moves a cluster network into a different compartment within the same tenancy. For
+     * information about moving resources between compartments, see
+     * [Moving Resources to a Different Compartment](https://docs.cloud.oracle.com/iaas/Content/Identity/Tasks/managingcompartments.htm#moveRes).
+     * <p>
+     * When you move a cluster network to a different compartment, associated resources such as the instances
+     * in the cluster network, boot volumes, and VNICs are not moved.
      *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
@@ -94,7 +99,9 @@ public interface ComputeManagement extends AutoCloseable {
             ChangeInstancePoolCompartmentRequest request);
 
     /**
-     * Create a cluster network.
+     * Creates a cluster network. For more information about cluster networks, see
+     * [Managing Cluster Networks](https://docs.cloud.oracle.com/iaas/Content/Compute/Tasks/managingclusternetworks.htm).
+     *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
      * @throws BmcException when an error occurs.
@@ -139,7 +146,7 @@ public interface ComputeManagement extends AutoCloseable {
     DetachLoadBalancerResponse detachLoadBalancer(DetachLoadBalancerRequest request);
 
     /**
-     * Gets the specified cluster network
+     * Gets information about the specified cluster network.
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
      * @throws BmcException when an error occurs.
@@ -190,7 +197,7 @@ public interface ComputeManagement extends AutoCloseable {
             LaunchInstanceConfigurationRequest request);
 
     /**
-     * List the instances in the specified cluster network.
+     * Lists the instances in the specified cluster network.
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
      * @throws BmcException when an error occurs.
@@ -274,7 +281,11 @@ public interface ComputeManagement extends AutoCloseable {
     StopInstancePoolResponse stopInstancePool(StopInstancePoolRequest request);
 
     /**
-     * Terminate the specified cluster network.
+     * Terminates the specified cluster network.
+     * <p>
+     * When you delete a cluster network, all of its resources are permanently deleted,
+     * including associated instances and instance pools.
+     *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
      * @throws BmcException when an error occurs.
@@ -290,8 +301,7 @@ public interface ComputeManagement extends AutoCloseable {
     TerminateInstancePoolResponse terminateInstancePool(TerminateInstancePoolRequest request);
 
     /**
-     * Update the specified cluster network.
-     * The OCID of the cluster network remains the same.
+     * Updates the specified cluster network. The OCID of the cluster network remains the same.
      *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation