+++ bmc-core/src/main/java/com/oracle/bmc/core/model/UpdateClusterNetworkDetails.java
@@ -4,7 +4,7 @@
 package com.oracle.bmc.core.model;
 
 /**
- * The data to update a cluster network
+ * The data to update a cluster network.
  * <br/>
  * Note: Objects should always be created or deserialized using the {@link Builder}. This model distinguishes fields
  * that are {@code null} because they are unset from fields that are explicitly set to {@code null}. This is done in
@@ -92,7 +92,9 @@ public class UpdateClusterNetworkDetails {
     java.util.Map<String, java.util.Map<String, Object>> definedTags;
 
     /**
-     * The user-friendly name.  Does not have to be unique.
+     * A user-friendly name. Does not have to be unique, and it's changeable. Avoid
+     * entering confidential information.
+     *
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("displayName")
     String displayName;