+++ bmc-core/src/main/java/com/oracle/bmc/core/model/UpdateVirtualCircuitDetails.java
@@ -52,6 +52,15 @@ public class UpdateVirtualCircuitDetails {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("customerAsn")
+        private Long customerAsn;
+
+        public Builder customerAsn(Long customerAsn) {
+            this.customerAsn = customerAsn;
+            this.__explicitlySet__.add("customerAsn");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("definedTags")
         private java.util.Map<String, java.util.Map<String, Object>> definedTags;
 
@@ -125,6 +134,7 @@ public class UpdateVirtualCircuitDetails {
                             bandwidthShapeName,
                             crossConnectMappings,
                             customerBgpAsn,
+                            customerAsn,
                             definedTags,
                             displayName,
                             freeformTags,
@@ -142,6 +152,7 @@ public class UpdateVirtualCircuitDetails {
                     bandwidthShapeName(o.getBandwidthShapeName())
                             .crossConnectMappings(o.getCrossConnectMappings())
                             .customerBgpAsn(o.getCustomerBgpAsn())
+                            .customerAsn(o.getCustomerAsn())
                             .definedTags(o.getDefinedTags())
                             .displayName(o.getDisplayName())
                             .freeformTags(o.getFreeformTags())
@@ -185,6 +196,14 @@ public class UpdateVirtualCircuitDetails {
     java.util.List<CrossConnectMapping> crossConnectMappings;
 
     /**
+     * Deprecated. Instead use `customerAsn`.
+     * If you specify values for both, the request will be rejected.
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("customerBgpAsn")
+    Integer customerBgpAsn;
+
+    /**
      * The BGP ASN of the network at the other end of the BGP
      * session from Oracle.
      * <p>
@@ -195,10 +214,12 @@ public class UpdateVirtualCircuitDetails {
      * If the BGP session is from the provider's edge router to Oracle, the
      * required value is the provider's ASN, and it can be updated only
      * by the provider.
+     * <p>
+     * Can be a 2-byte or 4-byte ASN. Uses \"asplain\" format.
      *
      **/
-    @com.fasterxml.jackson.annotation.JsonProperty("customerBgpAsn")
-    Integer customerBgpAsn;
+    @com.fasterxml.jackson.annotation.JsonProperty("customerAsn")
+    Long customerAsn;
 
     /**
      * Defined tags for this resource. Each key is predefined and scoped to a